Removed duplicated tests
